+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## TBD
 ### Changed
 * Refactored general purpose Hive metastore code to leverage hcommon-hive-metastore libraries. See [#72](https://github.com/HotelsDotCom/circus-train/issues/72).
+### Fixed
+* Avro schemas were not being replicated when a avro.schema.url without a scheme was specified. See [#74](https://github.com/HotelsDotCom/circus-train/issues/74)
 
 # 11.5.1 - 2018-05-24
 ### Fixed